requirements.txt
fastapi~=0.115.12
pymongo~=4.13.0
python-dotenv~=1.1.0
uvicorn~=0.34.2
requests~=2.32.3
apscheduler~=3.11.0

schemas~=0.7.1
auth~=0.5.3
pydantic~=2.11.5
db~=0.1.1
jose~=1.0.0
python-jose~=3.5.0